+++ templates/notice-flush-rewrite-rules.php
@@ -1,6 +1,6 @@
-<p><?= __('ACF Multilingual detected a change in your settings and needs to flush your rewrite rules.', 'acfml') ?></p>
+<p><?= wp_sprintf( __('%s Your settings have changed. You neeed to flush your rewrite rules and re-index some data.', 'acfml'), '[ACFML]') ?></p>
 <form method="POST">
   <?php wp_nonce_field('acfml_flush_rewrite_rules', '_acfml_nonce'); ?>
-  <input type="submit" class="button" value="<?= __('Flush Rewrite Rules Now', 'acfml') ?>">
+  <input type="submit" class="button" value="<?= __('Flush rewrite rules and process posts', 'acfml') ?>">
 </form>
 <p></p>