+++ templates/notice-flush-rewrite-rules.php
@@ -1,6 +1,5 @@
-<p><?= wp_sprintf( __('%s Your settings have changed. You neeed to flush your rewrite rules and re-index some data.', 'acfml'), '[ACFML]') ?></p>
+<p><?= __('[ACFML] Your settings have changed.', 'acfml') ?></p>
 <form method="POST">
   <?php wp_nonce_field('acfml_flush_rewrite_rules', '_acfml_nonce'); ?>
-  <input type="submit" class="button" value="<?= __('Flush rewrite rules and process posts', 'acfml') ?>">
+  <input type="submit" class="button" value="<?= __('Flush Rewrite Rules and Reprocess Posts', 'acfml') ?>">
 </form>
-<p></p>